# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Inventario de Redes",
        "description": "Control de creación, actualización, visualización y eliminación de registros de usuarios",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/devices": {
            "get": {
                "tags": [
                    "Devices"
                ],
                "summary": "( Muestra todos los dispositivos registrados )",
                "operationId": "fdbb50bf1d0489887b3b8c2e17c48923",
                "responses": {
                    "200": {
                        "description": "Datos recuperados exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Datos recuperados exitosamente"
                                        },
                                        "status": {
                                            "type": "number",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "anyOf": [
                                                    {
                                                        "$ref": "#/components/schemas/Device1Schema"
                                                    },
                                                    {
                                                        "$ref": "#/components/schemas/Device2Schema"
                                                    }
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/devices/register": {
            "post": {
                "tags": [
                    "Devices"
                ],
                "summary": "( Crea un nuevo dispositivo)",
                "operationId": "5b5203424fe9892c031803e181c9052d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Router "
                                    },
                                    "manufacturer": {
                                        "type": "string",
                                        "example": "tp-link"
                                    },
                                    "model": {
                                        "type": "string",
                                        "example": "AC64"
                                    },
                                    "serial": {
                                        "type": "string",
                                        "example": "SNR100000234567823ASD324"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "VIT-RT031"
                                    },
                                    "observation": {
                                        "type": "string",
                                        "example": "adquirido en 2018"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Router dual band 2.4GHz & 5Ghz"
                                    },
                                    "type_id": {
                                        "type": "number",
                                        "example": "1"
                                    },
                                    "status_id": {
                                        "type": "number",
                                        "example": "1"
                                    },
                                    "location_id": {
                                        "type": "number",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dispositivo Creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Registro creado"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "user": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "Router "
                                                },
                                                "manufacturer": {
                                                    "type": "string",
                                                    "example": "tp-link"
                                                },
                                                "model": {
                                                    "type": "string",
                                                    "example": "AC64"
                                                },
                                                "serial": {
                                                    "type": "string",
                                                    "example": "SNR100000234567823ASD324"
                                                },
                                                "code": {
                                                    "type": "string",
                                                    "example": "VIT-RT031"
                                                },
                                                "observation": {
                                                    "type": "string",
                                                    "example": "adquirido en 2018"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Router dual band 2.4GHz & 5Ghz"
                                                },
                                                "status_id": {
                                                    "type": "number",
                                                    "example": "1"
                                                },
                                                "location_id": {
                                                    "type": "number",
                                                    "example": "1"
                                                },
                                                "type_id": {
                                                    "type": "number",
                                                    "example": "1"
                                                },
                                                "created_at": {
                                                    "type": "number",
                                                    "example": "2023-06-10T22:27:14.000000"
                                                },
                                                "updated_at": {
                                                    "type": "number",
                                                    "example": "2023-06-10T22:27:14.000000"
                                                },
                                                "id": {
                                                    "type": "number",
                                                    "example": "2"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}